requirements.txt
torch
python-telegram-bot